added 403 forbidden code for update and delete

+++ Amit/AmitCartV3NewTcs/AmitCartV3NewTcs/src/main/java/com/example/AmitCartV3NewTcs/service/SellerService.java
@@ -25,6 +25,27 @@ public class SellerService {
 	
 	@Autowired
 	CategoryRepo categoryRepo;
+
+	//403 Forbidden.. update and delete api if seller is not the one who added the product..
+	//in Art project, requirement is that delete api should be accesed by Owner and only creator of art can delete api otherwise
+	//403 Forbidden.. so getArtById then check if user id in artFromDb has is same as user id from security context..
+	//if yes then deleteById otherwise return 403 Forbidden
+
+	//here it can be like delete api should be accesed by seller and onlyu seller who added the product should be able to delete
+	//Product has seller.. so find producyBy id then check if userId in prodcutFromDb == userId from securitycontext
+	//if yes then delete by Id and if not return 403 forbidden..
+	//but we are doing deleteBySellerUserIdAndProductId..no actually in seller api we are simply doing deleteById
+
+	//here in requirement nothing says.. it only says that if customer tries to access seller api.. give 403
+	//it does not say that if seller did not add the product then return 403.. but it should be like this only..
+
+	//this wa requirement in art that while updating or deleting if not found then return not found
+
+	//This is written for get api.. findByCartUserUserIdAndProductProductId
+	//same way if required, can give forbidden for consumer api when updating or deleting from cart
+	//but currently given deleteByUserUserIdAndProductProductId.. so use this..
+	//actually in json passing product here but product is added by seller in db not by consumer..
+	//consumer only adds in cart.. so here with product id can not serach that belongs to which consumer
 	
 	public ResponseEntity<Object> findAllProductOwnedBySeller(){
 		UserModel userModel = (UserModel) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
@@ -92,16 +113,26 @@ public class SellerService {
 //				productFromDb=productRepo.save(product);
 //				return ResponseEntity.status(HttpServletResponse.SC_OK).body(productFromDb);
 //			}
-//			else {
-				Category categoryFromDb = categoryRepo.findByCategoryName(product.getCategory().getCategoryName());
-				if(categoryFromDb == null) {
-					categoryFromDb = categoryRepo.save(product.getCategory());
-				}
-				product.setCategory(categoryFromDb);
-				product.setSeller(userModel);//create relationship between product and seller
-				Product productFromDb = productRepo.save(product);
-				return ResponseEntity.status(HttpServletResponse.SC_CREATED).body(productFromDb);
-//			}
+
+			Product productFromDb=productRepo.findById(product.getProductId()).orElse(null);
+			if(productFromDb!=null) {
+				//productFromDb=productRepo.save(product);
+				return ResponseEntity.status(HttpServletResponse.SC_NOT_FOUND).body("Product Not Found");
+				//this wa requirement in art that while updating or deleting if not found then return not found
+			}
+			if (product.getSeller().getUserId() != userModel.getUserId()){
+				return ResponseEntity.status(HttpServletResponse.SC_FORBIDDEN).body("You don't have permission!!");
+			}
+
+			Category categoryFromDb = categoryRepo.findByCategoryName(product.getCategory().getCategoryName());
+			if(categoryFromDb == null) {
+				categoryFromDb = categoryRepo.save(product.getCategory());
+			}
+			product.setCategory(categoryFromDb);
+			product.setSeller(userModel);//create relationship between product and seller
+			productFromDb = productRepo.save(product);
+			return ResponseEntity.status(HttpServletResponse.SC_CREATED).body(productFromDb);
+
 		}
 		catch (Exception e) {
 			e.printStackTrace();
@@ -110,11 +141,19 @@ public class SellerService {
 	}
 	
 	public ResponseEntity<Object> deleteProductBySellerId(int  productId){
+
 		try {
 			UserModel userModel = (UserModel) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-			Product productFromDb= productRepo.findByProductIdAndSellerUserId(productId, userModel.getUserId());
+//			Product productFromDb= productRepo.findByProductIdAndSellerUserId(productId, userModel.getUserId());
+//			if(productFromDb==null) {
+//				return ResponseEntity.status(HttpServletResponse.SC_NOT_FOUND).body("Product not owned by seller so can not delete");
+//			}
+			Product productFromDb= productRepo.findById(productId).orElse(null);
 			if(productFromDb==null) {
-				return ResponseEntity.status(HttpServletResponse.SC_NOT_FOUND).body("Product not owned by seller so can not delete");
+				return ResponseEntity.status(HttpServletResponse.SC_NOT_FOUND).body("Product not found");
+			}
+			if (productFromDb.getSeller().getUserId()!= userModel.getUserId()){
+				return ResponseEntity.status(HttpServletResponse.SC_FORBIDDEN).body("Product not owned By the logged in seller");
 			}
 			productRepo.deleteById(productId);
 			//we have already put check that no user other than seller can access this in security config
